requirements.txt
fastapi
uvicorn
google-generativeai
pymongo